refactor: ♻️ standardizes filenames to kebab-case
Renames feature and component files from snake_case to kebab-case. This change improves consistency across the codebase and aligns with common file naming conventions for modules in web development. All affected import paths are updated accordingly.
