+++ src/app/repos/page.tsx
@@ -1,6 +1,6 @@
 import ReactGithubCalendar from '@/components/calendar';
 import Layout from '@/components/layout';
-import RepoList from '@/features/repos/repo_list';
+import RepoList from '@/features/repos/repo-list';
 import { readRepositories } from '@/lib/repos';
 
 export default async function Portfolio() {
