+++ src/test/lib/simple-icons.test.ts
@@ -1,6 +1,6 @@
 import { describe, expect, it } from 'vitest';
 
-import { getSimpleIcon } from '@/features/repos/simple_icons';
+import { getSimpleIcon } from '@/features/repos/simple-icons';
 
 describe('getSimpleIcon', () => {
   it('returns icon info when spec exists', () => {
